Agents/strands-agents-with-memory/game-analyst-agent-with-memory/requirements.txt
strands-agents
strands-agents-tools
uv
uvicorn
boto3
bedrock-agentcore
bedrock-agentcore-starter-toolkit
aws-opentelemetry-distro